feat: support sqlalchemy metadata reflection (#55)
The main drawback: I dont see a way to rigorously support the `pk_`
columns of `IMPORTED KEYS`. And/or generally a bunch of the snowflake
columns, I am hardcoding values for.

From a full compatibility perspective, it's not great. But
realistically, it will never be perfect, and it **seems** to be
irrelevant to SQLAlchemy

---------

Co-authored-by: Oliver Mannion <125105+tekumara@users.noreply.github.com>