feat: cast string literals to timestamp in dateadd (#72)
Snowflake -again- seems to cast string literal in `DATEADD` to
`TIMESTAMP_LTZ` or `TIMESTAMP_NTZ` depending on `TIMESTAMP_TYPE_MAPPING`
session parameter.

```sql
select
  dateadd('day', 3, '2023-04-02') as c1, SYSTEM$TYPEOF(c1),
  dateadd('week', 3, '2023-04-02') as c3, SYSTEM$TYPEOF(c3),
  dateadd('month', 3, '2023-04-02') as c5, SYSTEM$TYPEOF(c5),
  dateadd('year', 3, '2023-04-02') as c7, SYSTEM$TYPEOF(c7),
;
```

|              C1               |   SYSTEM$TYPEOF(C1)    |
|-------------------------------|------------------------|
| 2023-04-05 00:00:00.000 +0000 | TIMESTAMP_LTZ(9)[SB16] |

|              C3               |   SYSTEM$TYPEOF(C3)    |
|-------------------------------|------------------------|
| 2023-04-23 00:00:00.000 +0000 | TIMESTAMP_LTZ(9)[SB16] |

|              C5               |   SYSTEM$TYPEOF(C5)    |
|-------------------------------|------------------------|
| 2023-07-02 00:00:00.000 +0000 | TIMESTAMP_LTZ(9)[SB16] |

|              C7               |   SYSTEM$TYPEOF(C7)    |
|-------------------------------|------------------------|
| 2026-04-02 00:00:00.000 +0000 | TIMESTAMP_LTZ(9)[SB16] |


Here seems to me we can cast the operand to `TIMESTAMP` if it is a
string literal to achieve similar behaviour.

But I'm a little bit concerned on `TIMESTAMP` as DuckDB is quite picky
operations between `TIMESTAMP` and `TIMESTAMP WITH TIMEZONE`. Like;
`CURRENT_TIMESTAMP` returns `TIMESTAMP WITH TIMEZONE` and `DATE_DIFF`
expects either operands to be same typed. I saw that `to_timestamp`
transformation casts to `TIMESTAMP` and went with it.

---------

Co-authored-by: Oliver Mannion <125105+tekumara@users.noreply.github.com>