feat: Adds MERGE INTO transform (#109)
This commit adds the ability to convert snowflakes [MERGE
INTO](https://docs.snowflake.com/en/sql-reference/sql/merge)
functionality into a functional equivalent implementation in duckdb. To
do this we need to break apart the WHEN [NOT] MATCHED syntax into
separate statements to be executed independently.

This commit **only adds the transform**, there is more refactoring
required in `fakes.py` in order to handle a transform that transforms a
single expression into multiple expressions so it cannot be used yet. I
have a change baking for adding it into `fakes` but wanted to separate
PRs.

Let me know if you need more tests or anything as its quite a
complicated feature.

---------

Co-authored-by: Oliver Mannion <125105+tekumara@users.noreply.github.com>