feat(server): basic sessionInfo login response (#222)
This fixes the .NET connector throwing an error when attempting to
connect as it expects a sessionInfo object on the data.

```
> dotnet run snowflake-test.cs
Unhandled exception. Snowflake.Data.Client.SnowflakeDbException (0x80004005): Error: Snowflake Internal Error: Unable to connect. Object reference not set to an instance of an object. SqlState: 08006, VendorCode: 270001, QueryId:
 ---> System.NullReferenceException: Object reference not set to an instance of an object.
   at Snowflake.Data.Core.SFSession.ProcessLoginResponse(LoginResponse authnResponse)
...
```

After this fix fakesnow works from C# and I'm able to successfully run
queries.

I thought it was safer to put an empty object to make this connector
work rather than populating its fields, but I'm happy to populate the
database, schema, warehouse and role if you think they should be in the
response.